Faster pip checks in mfc.sh, relative link for ___current___, and split mfc.py into multiple files
